Reorganize the code - rename package -add utils files and unit tests
